+++ background_scripts/options.js
@@ -65,7 +65,7 @@ var defaultSettings = {
   homedirectory: '',
   activehighlight: '#ff9632',
   highlight: '#ffff00',
-  completionengines: [],
+  completionengines: ['google', 'duckduckgo', 'wikipedia', 'amazon'],
   nextmatchpattern: '((?!first)(next|older|more|>|›|»|forward| )+)',
   previousmatchpattern: '((?!last)(prev(ious)?|newer|back|«|less|<|‹| )+)',
   barposition: 'top',